backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.4",
    "uvicorn>=0.40.0",
    "python-dotenv>=1.2.1",
    "pydantic>=2.12.5",
    "aiosqlite>=0.22.1",
    "google-genai>=1.62.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
